style(home page): add hover effect to the ticket card and remove unused imports
