+++ src/components/Home/HeroSection.jsx
@@ -28,44 +28,44 @@ function HeroSection() {
     };
 
     return (
-            <Slider {...settings} className='bg-white'>
-                {slideImages.map((img) => (
-                    <>
-                        <div
-                            key={img.id}
-                            style={{
-                                backgroundImage: `url(/Images/${img.imgSrc}.jpg)`,
-                            }}
-                            className='h-[90vh] bg-cover bg-center relative'
-                        >
-                            <div className=' flex flex-col justify-between absolute leading-loose bottom-[25%] text-center left-48 right-48 p-6 bg-opacity-50 bg-white bg-clip-padding'>
-                                <div className='space-y-4'>
-                                    <h2 className='text-lg md:text-xl lg:text-7xl leading-loose sm:text-md text-black font-extrabold font-atkinson'>
-                                        {t("hero.heading")}
-                                        <br />
-                                    </h2>
-                                    <p className='text-lg md:text-xl lg:text-3xl sm:text-md font-medium text-black leading-loose font-atkinson'>
-                                        {t("hero.text1")}
-                                    </p>
-                                    <p className='hidden sm:block text-lg md:text-xl lg:text-3xl font-medium  font-atkinson leading-loose'>
-                                        {t("hero.text2")}
-                                    </p>
-                                </div>
-                                <div className='px-auto pt-6 mx-auto'>
-                                    <Link href='/bookAnAppointment'>
-                                        <Button
-                                            color='teal'
-                                            transition={true}
-                                            buttonText={t("hero.bookingButton")}
-                                            buttonSize='fit'
-                                        />
-                                    </Link>
-                                </div>
+        <Slider {...settings} className='bg-white'>
+            {slideImages.map((img) => (
+                <>
+                    <div
+                        key={img.id}
+                        style={{
+                            backgroundImage: `url(/Images/${img.imgSrc}.jpg)`,
+                        }}
+                        className='h-[90vh] bg-cover bg-center relative'
+                    >
+                        <div className=' flex flex-col justify-between absolute leading-loose bottom-[25%] text-center left-48 right-48 p-6 bg-opacity-50 bg-white bg-clip-padding'>
+                            <div className='space-y-4'>
+                                <h2 className='text-lg md:text-xl lg:text-7xl leading-loose sm:text-md text-black font-extrabold font-atkinson'>
+                                    {t("hero.heading")}
+                                    <br />
+                                </h2>
+                                <p className='text-lg md:text-xl lg:text-3xl sm:text-md font-medium text-black leading-loose font-atkinson'>
+                                    {t("hero.text1")}
+                                </p>
+                                <p className='hidden sm:block text-lg md:text-xl lg:text-3xl font-medium  font-atkinson leading-loose'>
+                                    {t("hero.text2")}
+                                </p>
+                            </div>
+                            <div className='px-auto pt-6 mx-auto'>
+                                <Link href='/bookAnAppointment'>
+                                    <Button
+                                        color='teal'
+                                        transition={true}
+                                        buttonText={t("hero.bookingButton")}
+                                        buttonSize='fit'
+                                    />
+                                </Link>
                             </div>
                         </div>
-                    </>
-                ))}
-            </Slider>
+                    </div>
+                </>
+            ))}
+        </Slider>
 
         // -----------------------------------------------------------------------------------------
