+++ src/util/library.js
@@ -1,69 +1,45 @@
 export const slideImages = [
     {
         imgSrc: "2",
-        id: "8",
+        id: "1",
     },
     {
         imgSrc: "4",
-        id: "13",
+        id: "2",
     },
-    // {
-    //     imgSrc: "11",
-    //     id: "1",
-    // },
-    // {
-    //     imgSrc: "12",
-    //     id: "2",
-    // },
-    // {
-    //     imgSrc: "5",
-    //     id: "3",
-    // },
     {
         imgSrc: "6",
-        id: "4",
+        id: "3",
     },
-    // {
-    //     imgSrc: "7",
-    //     id: "5",
-    // },
     {
         imgSrc: "8",
-        id: "6",
+        id: "4",
     },
     {
         imgSrc: "13",
-        id: "7",
+        id: "5",
     },
 
     {
         imgSrc: "1",
-        id: "9",
+        id: "6",
+    },
+    {
+        imgSrc: "17",
+        id: "7",
+    },
+    
+    {
+        imgSrc: "14",
+        id: "8",
     },
     // {
     //     imgSrc: "18",
-    //     id: "10",
-    // },
-    // {
-    //     imgSrc: "25",
-    //     id: "11",
-    // },
-    {
-        imgSrc: "17",
-        id: "12",
-    },
-
-    {
-        imgSrc: "14",
-        id: "14",
-    },
-    // {
-    //     imgSrc: "30",
-    //     id: "15",
+    //     id: "9",
     // },
     // {
     //     imgSrc: "20",
-    //     id: "15",
+    //     id: "10",
     // },
 ];
